requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
openai==0.27.0
python-docx==0.8.11